requirements.txt
gitpython
tk